More @Bind -> @Binds renaming artifacts ------------- Created by MOE: https://github.com/google/moe MOE_MIGRATED_REVID=124872232
