+++ compiler/src/it/functional-tests/src/main/java/test/binds/Foo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package producerstest.bind;
+package test.binds;
 
 /**
  * This is the type that will be bound.  We throw in generics just to complicate the test.
