+++ compiler/src/it/functional-tests/src/main/java/test/binds/FooOfStrings.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package test.bind;
+package test.binds;
 
-/**
- * This is the type that will be bound.  We throw in generics just to complicate the test.
- */
-interface Foo<T> {}
+import javax.inject.Inject;
+
+final class FooOfStrings implements Foo<String> {
+  @Inject
+  FooOfStrings() {}
+}
