+++ compiler/src/it/functional-tests/src/main/java/test/binds/SimpleBindingModule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package test.bind;
+package test.binds;
 
 import dagger.Binds;
 import dagger.Module;
