+++ compiler/src/it/producers-functional-tests/src/main/java/producerstest/binds/SimpleBindingModule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package producerstest.bind;
+package producerstest.binds;
 
 import com.google.common.util.concurrent.MoreExecutors;
 import dagger.Binds;
