+++ compiler/src/it/producers-functional-tests/src/test/java/producerstest/binds/BindsProducersTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package producerstest.bind;
+package producerstest.binds;
 
 import com.google.common.collect.ImmutableMap;
 import dagger.producers.Produced;
